# apps/control-api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OAE Control API
  version: 1.0.0
  description: Canonical control-plane API for Odoo AI Edition.
servers:
  - url: http://localhost:4000
paths:
  /api/v1/builds:
    post:
      summary: Create a build request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - workspace_id
                - instance_id
                - title
                - prompt
                - execution_mode
                - deployment_path
              properties:
                workspace_id:
                  type: string
                instance_id:
                  type: string
                title:
                  type: string
                prompt:
                  type: string
                attachments:
                  type: array
                  items:
                    type: string
                execution_mode:
                  type: string
                  enum: [managed_api, chatgpt_edge]
                deployment_path:
                  type: array
                  items:
                    type: string
                    enum: [dev, sandbox, staging, production]
                risk_tolerance:
                  type: string
                  enum: [low, medium, high, critical]
                auto_deploy_sandbox:
                  type: boolean
      responses:
        '201':
          description: Build accepted
  /api/v1/builds/{build_id}:
    get:
      summary: Fetch build detail
      parameters:
        - in: path
          name: build_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Build detail
  /api/v1/builds/{build_id}/stream:
    get:
      summary: Stream build events
      parameters:
        - in: path
          name: build_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SSE stream
...
  /api/v1/edge/jobs/next:
    get:
      summary: Claim next edge job
      responses:
        '200':
          description: Job payload
        '204':
          description: No jobs available
...